Move /examples to /test/examples
Move usage examples to the testing folder
Update requires/references to match the new structure
